refactor: 관리자 예약 생성 api 클래스 분리
