+++ src/test/java/roomescape/controller/api/ReservationControllerTest.java
@@ -2,6 +2,8 @@ package roomescape.controller.api;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.hamcrest.Matchers.is;
+import static roomescape.util.Fixture.MEMBER_EMAIL;
+import static roomescape.util.Fixture.MEMBER_PASSWORD;
 import static roomescape.util.Fixture.PAST_TIME;
 import static roomescape.util.Fixture.TODAY;
 import static roomescape.util.Fixture.TOMORROW;
@@ -55,63 +57,6 @@ class ReservationControllerTest {
                 .when().post("/reservations")
                 .then().log().all()
                 .statusCode(201);
-    }
-
-    @Test
-    @DisplayName("관리자가 회원을 지정하여 예약을 생성한다.")
-    void createReservationByAdmin() {
-        final String adminToken = getAdminToken();
-        final Map<String, Object> params = new HashMap<>();
-        params.put("date", TOMORROW);
-        params.put("timeId", 1);
-        params.put("themeId", 1);
-        params.put("memberId", 1);
-
-        RestAssured.given().log().all()
-                .header("Cookie", adminToken)
-                .contentType(ContentType.JSON)
-                .body(params)
-                .when().post("/admin/reservations")
-                .then().log().all()
-                .statusCode(201);
-    }
-
-    @Test
-    @DisplayName("관리자 예약 생성 시, 관리자가 로그인한 경우가 아니면 예외가 발생한다.")
-    void createReservationByAdminWithNotAdmin() {
-        final String memberToken = getMemberToken();
-        final Map<String, Object> params = new HashMap<>();
-        params.put("date", TOMORROW);
-        params.put("timeId", 1);
-        params.put("themeId", 1);
-        params.put("memberId", 1);
-
-        RestAssured.given().log().all()
-                .header("Cookie", memberToken)
-                .contentType(ContentType.JSON)
-                .body(params)
-                .when().post("/admin/reservations")
-                .then().log().all()
-                .statusCode(401);
-    }
-
-    @Test
-    @DisplayName("관리자 예약 생성 시, memberId가 존재하지 않으면 예외가 발생한다.")
-    void createReservationByAdminWithNotFoundMemberId() {
-        final String adminToken = getAdminToken();
-        final Map<String, Object> params = new HashMap<>();
-        params.put("date", TOMORROW);
-        params.put("timeId", 1);
-        params.put("themeId", 1);
-        params.put("memberId", -1);
-
-        RestAssured.given().log().all()
-                .header("Cookie", adminToken)
-                .contentType(ContentType.JSON)
-                .body(params)
-                .when().post("/admin/reservations")
-                .then().log().all()
-                .statusCode(404);
     }
 
     @Test
@@ -377,17 +322,9 @@ class ReservationControllerTest {
     }
 
     private String getMemberToken() {
-        return getLoginToken("imjojo@gmail.com", "qwer");
-    }
-
-    private String getAdminToken() {
-        return getLoginToken("admin@gmail.com", "12345");
-    }
-
-    private String getLoginToken(final String email, final String password) {
         final Map<String, Object> loginParams = new HashMap<>();
-        loginParams.put("email", email);
-        loginParams.put("password", password);
+        loginParams.put("email", MEMBER_EMAIL);
+        loginParams.put("password", MEMBER_PASSWORD);
 
         return RestAssured.given().log().all()
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
